Unify iOS and macOS cross-compilers and switch to Clang 5.0 by default
Pull-request for branch users/somov/platforms

ref:63e47591e61b6635033d9b46e4b46442a07e3f64

+++ util/digest/multi_ut.cpp
@@ -14,11 +14,11 @@ class TMultiHashTest: public TTestBase {
 
 private:
     inline void TestStrInt() {
-        UNIT_ASSERT_EQUAL(MultiHash(TString("1234567"), static_cast<int>(123)), ULL(17038203285960021630));
+        UNIT_ASSERT_EQUAL(MultiHash(TString("1234567"), static_cast<int>(123)), static_cast<size_t>(ULL(17038203285960021630)));
     }
 
     inline void TestIntStr() {
-        UNIT_ASSERT_EQUAL(MultiHash(static_cast<int>(123), TString("1234567")), ULL(9973288649881090712));
+        UNIT_ASSERT_EQUAL(MultiHash(static_cast<int>(123), TString("1234567")), static_cast<size_t>(ULL(9973288649881090712)));
     }
 
     inline void TestSimpleCollision() {
